FEAT: Achieve BVH used to speed up rendering

+++ chrray-lib/include/light.h
@@ -11,7 +11,7 @@ public:
     virtual ~light() = default;
     virtual euclidean_coordinate direction(
         const euclidean_coordinate& hit_point) const = 0;
-    virtual double distance(const euclidean_coordinate& hit_point) const = 0;
+    virtual float distance(const euclidean_coordinate& hit_point) const = 0;
     virtual color intensity(const euclidean_coordinate& hit_point) const = 0;
     virtual ray shadow_ray(const euclidean_coordinate& hit_point) const = 0;
 };
@@ -21,22 +21,22 @@ public:
     point_light(
         const euclidean_coordinate& position,
         const color& intensity,
-        double constant_atten = 1.0,
-        double linear_atten = 0.0,
-        double quad_atten = 0.0);
+        float constant_atten = 1.0f,
+        float linear_atten = 0.0f,
+        float quad_atten = 0.0f);
 
     euclidean_coordinate direction(
         const euclidean_coordinate& hit_point) const override;
-    double distance(const euclidean_coordinate& hit_point) const override;
+    float distance(const euclidean_coordinate& hit_point) const override;
     color intensity(const euclidean_coordinate& hit_point) const override;
     ray shadow_ray(const euclidean_coordinate& hit_point) const override;
 
 private:
     euclidean_coordinate position_;
     color intensity_;
-    double const_atten_;
-    double lin_atten_;
-    double quad_atten_;
+    float const_atten_;
+    float lin_atten_;
+    float quad_atten_;
 };
 
 class directional_light : public light {
@@ -46,7 +46,7 @@ public:
 
     euclidean_coordinate direction(
         const euclidean_coordinate& hit_point) const override;
-    double distance(const euclidean_coordinate& hit_point) const override;
+    float distance(const euclidean_coordinate& hit_point) const override;
     color intensity(const euclidean_coordinate& hit_point) const override;
     ray shadow_ray(const euclidean_coordinate& hit_point) const override;
